requirements.txt
# API
fastapi
fastapi[standard]

celery
celery[redis]

# Scraping
Scrapy
scrapy-nodriver
websockets==13.1

# Data validation
pydantic

# Tests
pytest
